Rename cm.py to colormaps.py.
Closes #1476.
